Implement the missions data correctly and correct any errors
