+++ src/Pages/missionPage.js
@@ -1,9 +1,14 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
-import Mission from './missionComponent';
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import Mission from '../components/missionComponent';
+import { fetchMissions } from '../rocketsRedux/missions';
 
 const Missions = () => {
-  const missions = useSelector((state) => state);
+  const missions = useSelector((state) => state.missions);
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(fetchMissions());
+  }, []);
   return (
     <section>
       <table>
