+++ src/components/Rockets.js
@@ -1,10 +1,13 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { createActionForReserve, createActionForCancel } from '../rocketsRedux/fetchData';
+import fetchData, { createActionForReserve, createActionForCancel } from '../rocketsRedux/fetchData';
 
 const Rocket = () => {
   const dispatch = useDispatch();
-  const data = useSelector((state) => state);
+  useEffect(() => {
+    dispatch(fetchData());
+  }, []);
+  const dataRockets = useSelector((state) => state.rockets);
   const reserveHandler = (id) => {
     dispatch(createActionForReserve(id));
   };
@@ -15,9 +18,9 @@ const Rocket = () => {
 
   return (
     <section className="main-rocket-div">
-      {data.map((rocket) => (
+      {dataRockets.map((rocket) => (
         <div key={rocket.id} className="rocket-divider" id={rocket.id}>
-          <img className="rocket-images" src={rocket.flickr_images[0]} alt="Images" />
+          <img className="rocket-images" src={rocket.flickr_images} alt="Images" />
           <div>
             <h3>{rocket.rocket_name}</h3>
             <p>
