+++ src/index.js
@@ -16,3 +16,7 @@ root.render(
     </BrowserRouter>
   </React.StrictMode>,
 );
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
