+++ src/index.scss
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Inter&display=swap');
+// MISSIONS
 @import url('https://fonts.googleapis.com/css?family=Montserrat:thin,extra-light,light,100,200,300,400,500,600,700,800');
 @import url('https://fonts.googleapis.com/css?family=Roboto:thin,extra-light,light,100,200,300,400,500,600,700,800');
 
@@ -17,10 +18,6 @@ $azure: #18a2b8;
 * {
   margin: 0;
   padding: 0;
-}
-
-.App {
-  font-family: $roboto;
 }
 
 //NAVBAR
@@ -57,10 +54,6 @@ div.title {
   display: flex;
 }
 
-h1 {
-  margin-top: 25px;
-}
-
 .profile {
   border-left: 2px solid #000;
   padding-left: 10px;
@@ -70,11 +63,6 @@ h1 {
   width: 200px;
   height: 150px;
 }
-
-// #cancel {
-//   background-color: green;
-//   color: white;
-// }
 
 .main-rocket-div {
   display: flex;
@@ -104,6 +92,41 @@ button {
   padding: 6px 8px;
   background-color: #dfdfdf;
   border: 1px solid #d5d5d5;
+  transition: all 0.7s cubic-bezier(0.075, 0.82, 0.165, 1);
+
+  &:hover {
+    cursor: pointer;
+    box-shadow: rgba(0, 0, 0, 0.24) 0 3px 8px;
+    transform: scale(1.05);
+    transition: all 0.7s cubic-bezier(0.075, 0.82, 0.165, 1);
+  }
+
+  &:active {
+    transform: scale(1.01);
+  }
+}
+
+.reserve-text {
+  background-color: lightblue;
+  padding: 4px;
+  border-radius: 6px;
+  margin-right: 8px;
+}
+
+.myProfile {
+  display: flex;
+  justify-content: space-between;
+  padding: 0 20%;
+}
+
+//Missions
+
+.App {
+  font-family: $roboto;
+}
+
+h1 {
+  margin-top: 25px;
 }
 
 //MISSIONS
@@ -195,25 +218,13 @@ div.myMissions {
     transform: scale(1.05);
     transition: all 0.7s cubic-bezier(0.075, 0.82, 0.165, 1);
   }
-
-  &:active {
-    transform: scale(1.01);
-  }
-}
-
-.reserve-text {
-  background-color: lightblue;
-  padding: 4px;
-  border-radius: 6px;
-  margin-right: 8px;
-}
-
-.myProfile {
-  display: flex;
-  justify-content: space-between;
-  padding: 0 20%;
 }
 
 .joined {
   margin-left: 13px;
+}
+
+.profiles {
+  display: flex;
+  justify-content: center;
 }
