+++ src/rocketsRedux/store.js
@@ -1,7 +1,14 @@
-import { legacy_createStore as createStore, applyMiddleware } from 'redux';
+import { configureStore } from '@reduxjs/toolkit';
+import { applyMiddleware, combineReducers } from 'redux';
 import thunk from 'redux-thunk';
+import logger from 'redux-logger';
 import { myReducer } from './fetchData';
+import { missionsReducer } from './missions';
 
-const store = createStore(myReducer, applyMiddleware(thunk));
+const rootReducer = combineReducers({
+  rockets: myReducer,
+  missions: missionsReducer,
+});
+const store = configureStore({ reducer: rootReducer }, applyMiddleware(thunk, logger));
 
 export default store;
